+++ src/main/java/com/project200/undabang/member/service/MemberBlockCommandService.java
@@ -6,6 +6,5 @@ import java.util.UUID;
 
 public interface MemberBlockCommandService {
     CreateMemberBlockResponse createMemberBlock(UUID blockMemberId);
-
     void unBlockMember(UUID blockMemberId);
 }
