+++ src/main/java/com/project200/undabang/member/service/impl/MemberBlockCommandServiceImpl.java
@@ -1,9 +1,5 @@
 package com.project200.undabang.member.service.impl;
 
-import com.project200.undabang.chat.dto.event.ChatroomMemberStatusEvent;
-import com.project200.undabang.chat.entity.ChatroomMember;
-import com.project200.undabang.chat.repository.ChatroomMemberRepository;
-import com.project200.undabang.chat.repository.ChatroomRepository;
 import com.project200.undabang.common.context.UserContextHolder;
 import com.project200.undabang.common.web.exception.CustomException;
 import com.project200.undabang.common.web.exception.ErrorCode;
@@ -20,7 +16,6 @@ import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Optional;
 import java.util.UUID;
 
 @Slf4j
@@ -45,14 +40,7 @@ public class MemberBlockCommandServiceImpl implements MemberBlockCommandService 
         Member member = getMember(memberId);
         Member blockedMember = getMember(blockMemberId);
 
-        Optional<MemberBlock> optionalMemberBlock = memberBlockRepository.findByBlockerAndBlocked(member, blockedMember);
-
-        if (optionalMemberBlock.isPresent()) {
-            return handleMemberBlockExist(optionalMemberBlock.get());
-        }
-
-        MemberBlock savedMemberBlock = memberBlockRepository.save(MemberBlock.of(member, blockedMember));
-
+        MemberBlock savedMemberBlock = validateAndSaveMemberBlock(member, blockedMember);
         eventPublisher.publishEvent(MemberBlockedEvent.of(blockedMember, member));
 
         return CreateMemberBlockResponse.of(savedMemberBlock.getId());
@@ -90,18 +78,25 @@ public class MemberBlockCommandServiceImpl implements MemberBlockCommandService 
     }
 
     /**
-     * 기존의 MemberBlock이 존재할 경우 해당 상태를 처리하고 응답을 반환하는 메서드.
-     * 존재할 경우 409 Duplicated
-     * 존재하지만, 삭제한 경우 deletedAt = null 업데이트
+     * 주어진 MemberBlock 객체가 이미 존재하는지 확인하고, 기존 차단 상태를 처리하거나 재활성화하는 메서드.
      */
-    private CreateMemberBlockResponse handleMemberBlockExist(MemberBlock memberBlock) {
-
+    private MemberBlock handleMemberBlockExist(MemberBlock memberBlock) {
         if (memberBlock.getMemberBlockDeletedAt() == null) {
             throw new CustomException(ErrorCode.MEMBER_BLOCK_DUPLICATED);
-        } else {
-            memberBlock.reBlock();
-            return CreateMemberBlockResponse.of(memberBlock.getId());
         }
+
+        // 차단 상태라면 다시 활성화해줌
+        memberBlock.reBlock();
+        return memberBlock;
+    }
+
+    /**
+     * 회원 차단 정보를 검증하고 저장하는 메서드. 차단 정보가 이미 존재하는 경우 이를 처리하거나, 새롭게 차단 정보를 생성하여 저장합니다.
+     */
+    private MemberBlock validateAndSaveMemberBlock(Member member, Member blockedMember) {
+        return memberBlockRepository.findByBlockerAndBlocked(member, blockedMember)
+                .map(this::handleMemberBlockExist) // 이미 존재할 경우 예외 반환
+                .orElseGet(() -> memberBlockRepository.save(MemberBlock.of(member, blockedMember)));
     }
 
     /**
